type fix

+++ Arduino/esp32_solar_controller/beep.ino
@@ -1,4 +1,4 @@
-void beep_helper(const int freq, const int tdelay)
+void beep_helper(const uint16_t freq, const uint16_t tdelay)
 {
   ledcWriteTone(0, freq);
   delay(tdelay);
